# openapi.yml
openapi: 3.0.0
servers:
  - url: https://api.wp1.openzim.org/v1
    description: Production server HTTPS
  - url: http://api.wp1.openzim.org/v1
    description: Production server HTTP
info:
  description: 'Wikipedia 1.0 is an editing project on Wikipedia. It includes a bot, called the WP 1.0 bot, and a frontend web interface for that bot which lives at https://wp1.openzim.org. This document describes the server that acts as the API backend for that frontend server. Generally speaking, it contains only features which are useful to that frontend. However it is documented here and made available without restriction in the hopes that it may be useful.'
  version: 'v1'
  title: 'WP 1.0 Frontend'
  contact:
    name: 'Travis Briggs'
    email: 'audiodude@gmail.com'
  license:
    name: 'GPL v2 or greater'
    url: 'https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html'
...
paths:
...
  /zim/confirm-email:
    get:
      summary: 'Confirm an email address for ZIM file generation notifications.'
      description: 'The link containing this URL is emailed to the address given when a recurring ZIM schedule is created. The response is an HTML page intended to be viewed in a browser, not JSON.'
      operationId: confirmZimEmail
      tags:
        - ZIM
      responses:
        '200':
          description: 'The email address was confirmed.'
          content:
            text/html:
              schema:
                type: string
        '400':
          description: 'The token parameter was missing, or the confirmation failed.'
          content:
            text/html:
              schema:
                type: string
        '404':
          description: 'The token is invalid, expired, or has already been used.'
          content:
            text/html:
              schema:
                type: string
      parameters:
        - name: token
          in: query
          required: true
          description: 'The confirmation token from the email.'
          schema:
            type: string
  /zim/unsubscribe-email:
    get:
      summary: 'Unsubscribe from ZIM file generation notifications, using a confirmation token.'
      description: 'The response is an HTML page intended to be viewed in a browser, not JSON.'
      operationId: unsubscribeZimEmail
      tags:
        - ZIM
      responses:
        '200':
          description: 'The email address was unsubscribed.'
          content:
            text/html:
              schema:
                type: string
        '400':
          description: 'The token parameter was missing, or the unsubscribe failed.'
          content:
            text/html:
              schema:
                type: string
        '404':
          description: 'The token is invalid, expired, or has already been used.'
          content:
            text/html:
              schema:
                type: string
      parameters:
        - name: token
          in: query
          required: true
          description: 'The token from the email.'
          schema:
            type: string
  /zim/unsubscribe-notification:
    get:
      summary: 'Unsubscribe from ZIM file generation notifications using an emailed capability.'
      description: 'No login is required. The signed token is bound to the schedule and its current recipient email. Changed recipients and SESSION_SECRET_KEY rotation invalidate links. Old schedule_id-only links no longer authorize unsubscribe. The separate confirmation-token email workflow is unchanged. The response is an HTML page, not JSON.'
      operationId: unsubscribeZimNotification
      tags:
        - ZIM
      responses:
        '200':
          description: 'The email address was unsubscribed.'
          content:
            text/html:
              schema:
                type: string
        '400':
          description: 'The token parameter was missing.'
          content:
            text/html:
              schema:
                type: string
        '404':
          description: 'The token is invalid, the schedule or recipient no longer matches, or the recipient has already unsubscribed.'
          content:
            text/html:
              schema:
                type: string
      parameters:
        - name: token
          in: query
          required: true
          description: 'Purpose-separated signed capability from the notification email. Schedule IDs alone are not accepted.'
          schema:
            type: string
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: 'The session cookie set by /oauth/complete. Include credentials so this cookie is sent. All unsafe session-authenticated requests (POST, PUT, PATCH, DELETE, etc.) require an Origin exactly matching a configured CLIENT_DOMAINS HTTP(S) origin. Only when Origin is absent, a valid Referer with an exactly matching origin is accepted. Missing both headers, or an empty, null, malformed or untrusted Origin, produces 403 without mutation, even if Referer is trusted. Cookie-authenticated scripts must explicitly supply an allowed Origin. CLIENT_DOMAINS must be a non-empty list of literal origins without wildcards, patterns, userinfo, paths, queries or fragments. OAuth GET handshakes and token-authenticated webhook/email links are exempt.'
...